backend/requirements.txt
Flask
flask-cors
python-dotenv
langchain
langchain-groq
langchain-huggingface
langchain-community
chromadb
sentence-transformers
pypdf
werkzeug
